tribune_dictio/pyproject.toml
[project]
name = "ttsapi"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.11.*"
dependencies = [
    "bottle>=0.13.4",
    "chatterbox-tts>=0.1.4",
    "numpy>=1.24.0,<1.26.0",
    "librosa==0.11.0",
    "s3tokenizer",
    "torch==2.6.0",
    "torchaudio==2.6.0",
    "transformers==4.46.3",
    "diffusers==0.29.0",
    "resemble-perth==1.0.1",
    "conformer==0.3.2",
    "safetensors==0.5.3",
    "spacy-pkuseg",
    "pykakasi==2.3.0",
    "gradio==5.44.1",
    "onnx==1.18.0",
    "peft",
    "setuptools<70",
    "paste>=3.10.1",
]

[tool.uv.extra-build-dependencies]
pkuseg = ["numpy", "setuptools"]
